requirements.txt
pandas
jupyter
librosa
scikit-learn
tqdm
black
transformers
loguru
pyarrow
pytorch-lightning
torchmetrics
audiomentations[extras]
timm
scikit-image
azure-cognitiveservices-speech==1.25.0
boto3==1.26.65
datasets==2.9.0
google-cloud-texttospeech==2.14.1